+++ pkgs/dart_mcp_server/lib/src/server.dart
@@ -20,6 +20,7 @@ import 'mixins/analyzer.dart';
 import 'mixins/dash_cli.dart';
 import 'mixins/dtd.dart';
 import 'mixins/flutter_launcher.dart';
+import 'mixins/package_uri_reader.dart';
 import 'mixins/prompts.dart';
 import 'mixins/pub.dart';
 import 'mixins/pub_dev_search.dart';
@@ -45,7 +46,8 @@ final class DartMCPServer extends MCPServer
         DartToolingDaemonSupport,
         FlutterLauncherSupport,
         PromptsSupport,
-        DashPrompts
+        DashPrompts,
+        PackageUriSupport
     implements
         AnalyticsSupport,
         ProcessManagerSupport,