+++ pkgs/dart_mcp_server/CHANGELOG.md
@@ -6,6 +6,8 @@
 - Fix analyze tool handling of invalid roots.
 - Fix erroneous SDK version error messages when connecting to a VM Service
   instead of DTD URI.
+- Add a tool for reading package: URIs. Results are returned as embedded
+  resources or resource links (when reading a directory).
 
 # 0.1.1 (Dart SDK 3.10.0)